Create README.md
There are few law firms in the Houston area with the experience and reputation for handling complex family law matters that we have at Moffett Law Firm. Whether you have high-value assets, complicated property division concerns or are facing a contested custody matter, we have the skills to help you achieve your legal goals. We are experienced litigators and will fight aggressively for your rights in court when it benefits you to do so.